Upgrade to Fluture and Node version 12
