+++ .gitignore
@@ -1,2 +1,3 @@
+/coverage/
 /index.js
 /node_modules/
