+++ .travis.yml
@@ -1,3 +1,3 @@
 language: node_js
 node_js:
-  - "10"
+  - "12"
