+++ package.json
@@ -2,12 +2,13 @@
   "name": "fluture-hooks",
   "version": "1.0.2",
   "description": "Fantasy Land Monad and Alternative instances for Fluture's hook",
-  "main": "index",
-  "module": "index.mjs",
+  "type": "module",
+  "main": "index.cjs",
+  "module": "index.js",
   "scripts": {
-    "build": "rollup -c rollup.config.mjs",
+    "build": "rollup -c rollup.config.js",
     "release": "xyz --edit --script scripts/prerelease --increment",
-    "test": "node --experimental-modules --no-warnings --stack-trace-limit=100 ./test/index.mjs"
+    "test": "c8 --check-coverage --branches 100 --functions 100 --lines 100 --statements 100 node --experimental-modules --no-warnings ./node_modules/.bin/oletus"
   },
   "repository": {
     "type": "git",
@@ -23,13 +24,15 @@
   "author": "Aldwin Vlasblom <aldwin@avaq.it> (https://github.com/Avaq)",
   "license": "MIT",
   "dependencies": {
-    "callgebra": "^1.0.0"
+    "callgebra": "github:fluture-js/callgebra#avaq/maintenance"
   },
   "peerDependencies": {
-    "fluture": ">=11.0.0 <12.0.0"
+    "fluture": ">=12.0.0-beta.4 <13.0.0"
   },
   "devDependencies": {
-    "fluture": "^11.0.0",
+    "c8": "^5.0.4",
+    "fluture": "^12.0.0-beta.4",
+    "oletus": "^2.0.0",
     "rollup": "^1.21.4",
     "sanctuary-type-identifiers": "^2.0.1",
     "transcribe": "^1.1.1",
