--- rollup.config.mjs
@@ -1,15 +1,0 @@
-import pkg from './package.json';
-
-export default {
-  input: 'index.mjs',
-  external: (
-    Object.keys (pkg.peerDependencies)
-    .concat (Object.keys (pkg.dependencies))
-  ),
-  output: {
-    format: 'umd',
-    name: 'flutureHooks',
-    file: 'index.js',
-    interop: false
-  }
-};
